+++ exts/omni.osc/docs/CHANGELOG.md
@@ -2,7 +2,16 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.0] - 2022-09-12
+### Changed
+- The `On OSC Message` OmniGraph node now outputs a Bundle typed value rather than an Unknown typed value.
+    - Users can extract the "address" and the "arguments" of the OSC message with the `Extract Attribute` node.
+
+## [0.1.1] - 2022-09-12
+### Changed
+- Updated documentation.
 
 ## [0.1.0] - 2022-09-02
+### Added
 - Initial release.